code_indexer/requirements.txt
watchdog
tree-sitter
google-cloud-aiplatform
pinecone-client